fix(geometry,wasm): split geom_hash's segment fold out, tighten a doc comment — clear the module-size ratchet
The tolerance-clamp commit (3461df03) pushed two files over their
module_size_ratchet budgets:

- rust/geometry/src/geom_hash.rs grew to 423 lines (no allowlist row,
  over the 400-line house rule). Its tests were already split into
  geom_hash_tests.rs, so the seam here is different: add_mesh /
  add_mesh_with_origin / add_oriented_mesh (the per-segment triangle
  fold) move to a new child module, geom_accumulate.rs, matching the
  existing geom_closure.rs / geom_bounds.rs / geom_surface.rs pattern
  of `impl GeometryHasher` blocks split across sibling files that
  reach the struct's private fields directly. geom_hash.rs is now 272
  lines.

- rust/wasm-bindings/src/api/mod.rs grew 2 lines past its 1088-line
  allowlisted budget, from the doc comment the clamp fix added to
  setComputeGeometryHashes. Tightened that comment to the same
  substance in fewer lines, the same fix PR #2539 used for a different
  comment in this same file (its own, separate budget breach at the
  IfcAPI::new() panic-hook comment). The two PRs are not on the same
  branch and don't collide today, but both independently add lines
  against this file's shared 1088 budget — worth a human's attention
  at merge time.

Neither budget was raised.

Verified: module_size_ratchet passes; ifc-lite-geometry lib tests
609/609 (unchanged from baseline — no test moved or dropped, only
the reordered impls' single OrientVerdict import in
geom_hash_tests.rs); ifc-lite-wasm lib tests 34/34; clippy clean on
both crates.
